+++ packages/fields/src/index.ts
@@ -1,11 +1,9 @@
 export * from './fields';
 export * from './actions';
 export { default as CreateTemplatePartModal } from './components/create-template-part-modal';
-export { default as MediaEdit } from './components/media-edit';
 export type {
 	BasePostWithEmbeddedAuthor,
 	BasePost,
 	PostType,
 	Pattern,
-	MediaEditProps,
 } from './types';
