--- packages/media-utils/src/components/index.js
@@ -1,1 +1,0 @@
-export { default as MediaUpload } from './media-upload';
