+++ packages/media-utils/src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as MediaUpload } from './media-upload';
+export { default as MediaEdit } from './media-edit';
+export type { MediaEditProps } from './media-edit';
