+++ packages/media-utils/src/components/media-edit/style.scss
@@ -10,7 +10,7 @@ fieldset.fields-controls__media {
 	.fields-controls__media-row {
 		display: grid;
 		grid-template-columns: 1fr auto;
-		gap: $grid-unit-15;
+		gap: $grid-unit-10;
 		align-items: center;
 	}
