+++ packages/media-utils/src/style.scss
@@ -1,1 +1,2 @@
 @use "@wordpress/media-fields/build-style/style.css" as *;
+@use "./components/media-edit/style.scss" as *;
